# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OpsPilot AI customer operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MetricPoint {
  label: string;
  value: number;
}

export interface Dashboard {
  openTickets: number;
  urgentTickets: number;
  atRiskSlas: number;
  breachedSlas: number;
  resolvedToday: number;
  avgResponseMinutes: number;
  avgResolutionHours: number;
  aiAssistedTickets: number;
  volume: MetricPoint[];
  statusBreakdown: MetricPoint[];
  priorityBreakdown: MetricPoint[];
  agentWorkload: MetricPoint[];
}

export interface Customer {
  id: string;
  name: string;
  company: string;
  email: string;
  phone?: string;
  tier: string;
  status: string;
  openTickets: number;
  lifetimeValue: number;
  updatedAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 3 */
  email: string;
  phone?: string;
  tier: string;
  status?: string;
}

export interface TimelineEvent {
  id: string;
  label: string;
  detail: string;
  createdAt: string;
}

export type CustomerDetail = Customer & {
  notes: string;
  timeline: TimelineEvent[];
};

export interface Ticket {
  id: string;
  title: string;
  description?: string;
  customer: string;
  customerId?: string;
  status: string;
  priority: string;
  category: string;
  source: string;
  assignee: string;
  team: string;
  sla: string;
  createdAt: string;
  updatedAt: string;
}

export interface TicketInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  customer: string;
  priority: string;
  category: string;
  source?: string;
}

export interface TicketUpdate {
  status?: string;
  priority?: string;
  assignee?: string;
  team?: string;
  reply?: string;
  internalNote?: string;
}

export interface Message {
  id: string;
  author: string;
  body: string;
  kind: string;
  createdAt: string;
}

export type AiAnalysisClassification = {[key: string]: string};

export interface AiAnalysis {
  provider: string;
  summary: string[];
  confidence: number;
  recommendation: string;
  classification: AiAnalysisClassification;
  suggestedReply: string;
  reviewed: boolean;
}

export type TicketDetail = Ticket & {
  messages: Message[];
  ai: AiAnalysis;
};

export interface KnowledgeDocument {
  id: string;
  title: string;
  excerpt: string;
  category: string;
  status: string;
  author: string;
  updatedAt: string;
  views: number;
}

export interface KnowledgeInput {
  title: string;
  content: string;
  category: string;
}

export interface AiActionInput {
  action: string;
}

export interface Notification {
  id: string;
  title: string;
  detail: string;
  kind: string;
  read: boolean;
  createdAt: string;
}

export interface AuditLog {
  id: string;
  action: string;
  entity: string;
  actor: string;
  detail: string;
  createdAt: string;
}

export type ListTicketsParams = {
search?: string;
status?: string;
priority?: string;
category?: string;
};

export type ListCustomersParams = {
search?: string;
status?: string;
};

export type ListKnowledgeParams = {
search?: string;
};