# lib/api-zod/src/generated/types/aiAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OpsPilot AI customer operations API
 * OpenAPI spec version: 0.1.0
 */
import type { AiAnalysisClassification } from './aiAnalysisClassification';

export interface AiAnalysis {
  provider: string;
  summary: string[];
  confidence: number;
  recommendation: string;
  classification: AiAnalysisClassification;
  suggestedReply: string;
  reviewed: boolean;
}