# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OpsPilot AI customer operations API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: string;
  action: string;
  entity: string;
  actor: string;
  detail: string;
  createdAt: string;
}