# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OpsPilot AI customer operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: string;
  name: string;
  company: string;
  email: string;
  phone?: string;
  tier: string;
  status: string;
  openTickets: number;
  lifetimeValue: number;
  updatedAt: string;
}