# lib/api-zod/src/generated/types/customerDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OpsPilot AI customer operations API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from './customer';
import type { TimelineEvent } from './timelineEvent';

export type CustomerDetail = Customer & {
  notes: string;
  timeline: TimelineEvent[];
};